ketemaderesa@outlook.com
User Experience Improvement Program

Dear users, Greetings!

ITEL MOBILE LIMITED and its global affiliates (hereinafter referred to as "ITEL" or "we") are committed to providing users with products that are of excellent value and that offer a good user experience. To help us understand the needs of users, improve the quality of products and provide personalized services to users, we sincerely welcome you to join this program. Before you join this program, please read carefully and fully understand the content of the program (children should be accompanied by a legal guardian when reading). The content of this program is designed to help you understand how to join this program, related rules and purposes. If you choose to join the program, it is deemed that you have carefully read and agreed to the content of the program.

In order to continuously improve the product operation experience and running fluency, improve functional design, adjust the priority of function iteration, and promptly provide you with personalized services that enhance your user experience, with your express consent, we will carry out statistics and analysis on your usage of various functions of the product and some general information regarding your device as needed, including but not limited to the duration of use of each function, the number of page clicks, application activation and exit status, model information, product identification code, software version number, device information, etc. Specifically, for example, in order to determine the significance of each product function and then adjust our priority for research and development of the product function, we may collect the usage and duration of each product function; in order to provide you with a better user experience, we may obtain some of the key logs when you encounter program crashes and product restarts during your use of the product, and then fix the specific problem and improve the functional design; in order to be able to distinguish different devices and provide error-free and available upgrade services, we may judge and provide appropriate upgrade services based on the obtained data such as the software version number, product identification code, model information, and system version number.

We promise that we will strictly abide by the relevant laws and regulations of your country/region, and protect you and your personal privacy. We only collect relevant information for the purpose of improving or enhancing the product performance and user experience, and use such information for the improvement of product experience. For more information on how we protect your privacy, please refer to the Privacy Policy.

You can freely join or opt out of this program via "Settings - System - About Phone - Legal Information - User Experience Improvement Program". Selecting "On" means that you agree to join this program; selecting "Off" means that you quit the program.

 